dataset-manager/pyproject.toml
[tool.poetry]
name = "dataset-manager"
version = "0.1.0"
description = "Scalable backend for dataset management platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.4.0"
pydantic-settings = "^2.0.3"
cassandra-driver = "^3.29.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"
python-multipart = "^0.0.6"
pandas = "^2.1.0"
pyarrow = "^13.0.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = ">=0.23.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
